+++ src/player/channel.ts
@@ -2,7 +2,7 @@ import { AudioInsert, connectSerial } from "./connect";
 import { createControl } from "./signals";
 import { midiVelToGain } from "./volume";
 
-export type ChannelOptions = {
+export type ChannelConfig = {
   destination: AudioNode;
   volume: number;
   volumeToGain: (volume: number) => number;
@@ -29,18 +29,17 @@ export class Channel {
   #inserts?: (AudioNode | AudioInsert)[];
   #disconnect: () => void;
   #unsubscribe: () => void;
-  #options: Readonly<ChannelOptions>;
+  #config: Readonly<ChannelConfig>;
   #disconnected = false;
 
   constructor(
     public readonly context: BaseAudioContext,
-    options: Partial<ChannelOptions>
+    options?: Partial<ChannelConfig>
   ) {
-    this.#options = {
-      destination: context.destination,
-      volume: 100,
-      volumeToGain: midiVelToGain,
-      ...options,
+    this.#config = {
+      destination: options?.destination ?? context.destination,
+      volume: options?.volume ?? 100,
+      volumeToGain: options?.volumeToGain ?? midiVelToGain,
     };
 
     this.input = context.createGain();
@@ -49,14 +48,13 @@ export class Channel {
     this.#disconnect = connectSerial([
       this.input,
       this.#volume,
-      this.#options.destination,
+      this.#config.destination,
     ]);
 
-    const volume = createControl(options.volume ?? 100);
+    const volume = createControl(this.#config.volume);
     this.setVolume = volume.set;
-    const volumeToGain = options.volumeToGain ?? midiVelToGain;
     this.#unsubscribe = volume.subscribe((volume) => {
-      this.#volume.gain.value = volumeToGain(volume);
+      this.#volume.gain.value = this.#config.volumeToGain(volume);
     });
   }
 
@@ -71,7 +69,7 @@ export class Channel {
       this.input,
       ...this.#inserts,
       this.#volume,
-      this.#options.destination,
+      this.#config.destination,
     ]);
   }