+++ src/player/queued-player.ts
@@ -1,25 +1,62 @@
 import { SortedQueue } from "./sorted-queue";
-import {
-  InternalPlayer,
-  SampleOptions,
-  SampleStart,
-  SampleStop,
-} from "./types";
+import { InternalPlayer, SampleStart, SampleStop } from "./types";
 
 type SampleStartWithTime = SampleStart & { time: number };
+
+function compose<T>(a?: (x: T) => void, b?: (x: T) => void) {
+  return a && b
+    ? (x: T) => {
+        a(x);
+        b(x);
+      }
+    : a ?? b;
+}
+
+export type QueuedPlayerConfig = {
+  scheduleLookaheadMs: number;
+  scheduleIntervalMs: number;
+  onStart?: (sample: SampleStart) => void;
+  onEnded?: (sample: SampleStart) => void;
+};
+
+function getConfig(options: Partial<QueuedPlayerConfig>) {
+  const config: QueuedPlayerConfig = {
+    scheduleLookaheadMs: options.scheduleLookaheadMs ?? 200,
+    scheduleIntervalMs: options.scheduleIntervalMs ?? 50,
+    onStart: options.onStart,
+    onEnded: options.onEnded,
+  };
+
+  if (config.scheduleLookaheadMs < 1) {
+    throw Error("scheduleLookaheadMs must be greater than 0");
+  }
+  if (config.scheduleIntervalMs < 1) {
+    throw Error("scheduleIntervalMs must be greater than 0");
+  }
+  if (config.scheduleLookaheadMs < config.scheduleIntervalMs) {
+    throw Error("scheduleLookaheadMs must be greater than scheduleIntervalMs");
+  }
+
+  return config;
+}
 
 /**
  * A SamplePlayer that queues up samples to be played in the future.
  *
  * @private
  */
-
 export class QueuedPlayer implements InternalPlayer {
   private readonly player: InternalPlayer;
+  #config: QueuedPlayerConfig;
   #queue: SortedQueue<SampleStartWithTime>;
   #intervalId: NodeJS.Timeout | undefined;
 
-  public constructor(player: InternalPlayer, options: Partial<SampleOptions>) {
+  public constructor(
+    player: InternalPlayer,
+    options: Partial<QueuedPlayerConfig> = {}
+  ) {
+    this.#config = getConfig(options);
+
     this.#queue = new SortedQueue<SampleStartWithTime>(
       (a, b) => a.time - b.time
     );
@@ -42,14 +79,18 @@ export class QueuedPlayer implements InternalPlayer {
     const context = this.player.context;
     const now = context.currentTime;
     const startAt = sample.time ?? now;
-    if (startAt < now + 1) {
+    const lookAhead = this.#config.scheduleLookaheadMs / 1000;
+    sample.onStart = compose(sample.onStart, this.#config.onStart);
+    sample.onEnded = compose(sample.onEnded, this.#config.onEnded);
+
+    if (startAt < now + lookAhead) {
       return this.player.start(sample);
     }
     this.#queue.push({ ...sample, time: startAt });
 
     if (!this.#intervalId) {
       this.#intervalId = setInterval(() => {
-        const nextTick = context.currentTime + 1;
+        const nextTick = context.currentTime + lookAhead;
         while (this.#queue.size() && this.#queue.peek()!.time <= nextTick) {
           const sample = this.#queue.pop();
           if (sample) {
@@ -60,7 +101,7 @@ export class QueuedPlayer implements InternalPlayer {
           clearInterval(this.#intervalId!);
           this.#intervalId = undefined;
         }
-      }, 200);
+      }, this.#config.scheduleIntervalMs);
     }
 
     return (time?: number) => {