+++ src/player/sample-player.test.ts
@@ -1,23 +1,34 @@
-import { createAudioContextMock } from "../test/audio-context-test-helpers";
+import { createAudioContextMock } from "../test-helpers";
 import { SamplePlayer } from "./sample-player";
 
 describe("SamplePlayer", () => {
-  it("sample name takes preference over note", () => {
-    const context = createAudioContextMock();
-    const player = new SamplePlayer(context.destination, {});
-    player.buffers[60] = context.createBuffer(1, 1, 44100);
+  it("note name can be a midi note", () => {
+    const contextMock = createAudioContextMock();
+    const player = new SamplePlayer(contextMock.context, {});
+    player.buffers[60] = contextMock.createBuffer(1, 1, 44100);
 
     player.start({ note: 60 });
-    expect((context as any).bufferSources[0].buffer).toBe(player.buffers[60]);
+    expect((contextMock as any).bufferSources[0].buffer).toBe(
+      player.buffers[60]
+    );
   });
 
   it("sample name takes preference over note", () => {
-    const context = createAudioContextMock();
-    const player = new SamplePlayer(context.destination, {});
-    player.buffers[60] = context.createBuffer(1, 1, 44100);
-    player.buffers["A"] = context.createBuffer(1, 1, 44100);
+    const contextMock = createAudioContextMock();
+    const player = new SamplePlayer(contextMock.context, {});
+    player.buffers[60] = contextMock.createBuffer(1, 1, 44100);
+    player.buffers["A"] = contextMock.createBuffer(1, 1, 44100);
 
     player.start({ note: 60, name: "A" });
-    expect(context.bufferSources[0].buffer).toBe(player.buffers["A"]);
+    expect(contextMock.bufferSources[0].buffer).toBe(player.buffers["A"]);
+  });
+
+  it("calls onStart", () => {
+    const contextMock = createAudioContextMock();
+    const player = new SamplePlayer(contextMock.context, {});
+    player.buffers["A"] = contextMock.createBuffer(1, 1, 44100);
+    const onStart = jest.fn();
+    player.start({ note: 60, name: "A", onStart });
+    expect(onStart).toHaveBeenCalledWith({ note: 60, name: "A", onStart });
   });
 });